Export the dst/0 type from erlang.erl
The dst/0 type describes what types of values that it is possible to send messages to in Erlang. It would be useful to have one source of truth for this type when referring to it in other modules, since it is somewhat complex it is easy to miss acceptable values when defining it yourself.